requirements.txt
Flask==3.0.3
flask-socketio==5.3.6
mysql-connector-python==8.3.0
eventlet==0.36.1
Werkzeug==3.0.3
Flask-Mail==0.10.0
twilio==9.3.4
python-docx==1.1.2
openpyxl==3.1.5
